HTR segmentation, and GAN for datagen
An attempt to fit a pretrained character recognition model, to perform HTR in order to achieve OCR. Graffiteers are artists, it s hard to evaluate art!!! What is not so hard it to distinguish bad art from good art. However, for the HTR case, it is not about art.... It s about achieving an image caption recogntion character level with correct word and letter order for as much art as possible. The idea is to introduce noise characters or dump them, and reach the same OCR as a CRNN would achieve, and the chosen way was mainly manifested for two reason. There s too much noise on a random image, e.g. there was a graffiti on the shopkeeper's or another one in a wooden window cover. That was the first reason, in simple terms, the problem was not easy to determine. After determining that at least some of the quote lines would be readable by OCR, then it became overly convoluted and savage to find a descent way to train a CrNN, or even Rnn for such a wicked task! However, doing it character by character is not impossible for these same terms, yet it can t be stated that on its way that s better to use than edge Neural Nets. In addition, to the Segmentation, a general adverserial was created, to supply a CNN with sufficient training data, to maximize the output success rates, in the likelyhood of a thereof better alternative lack, which seems possible due to the nature of this problem. The datasets were found on kaggle, the letters are in greek, high quality images were used